Move legacy button into separate file
